+++ Code_projet/projet_echec_et_pat/projet_echec_et_pat.layout
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8" standalone="yes" ?>
+<CodeBlocks_layout_file>
+	<FileVersion major="1" minor="0" />
+	<ActiveTarget name="Debug" />
+</CodeBlocks_layout_file>
